requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
twilio==8.13.0
deep-translator==1.11.4
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2
pickle-mixin==1.0.2
smtplib==0.1.0